requirements.txt
openai
python-dotenv
fastapi
pydantic
requests
uvicorn
pillow